+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* Added basic usage example of the library
+
+
 ## [1.16.1] - 2023-11-07
 ### Fixed
 * Fixed typo in error message when no auth key is provided.
@@ -273,6 +278,7 @@ Version increased to avoid conflicts with old packages on PyPI.
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-python/compare/v1.16.1...HEAD
 [1.16.1]: https://github.com/DeepLcom/deepl-python/compare/v1.16.0...v1.16.1
 [1.16.0]: https://github.com/DeepLcom/deepl-python/compare/v1.15.0...v1.16.0
 [1.15.0]: https://github.com/DeepLcom/deepl-python/compare/v1.14.0...v1.15.0